+++ tests/QueryBuilderTest.php
@@ -2,6 +2,8 @@
 
 namespace Tests;
 
+use MisterIcy\SqlQueryBuilder\Expressions\DML\Join;
+use MisterIcy\SqlQueryBuilder\Operations\Eq;
 use MisterIcy\SqlQueryBuilder\Operations\Gt;
 use MisterIcy\SqlQueryBuilder\Operations\Gte;
 use MisterIcy\SqlQueryBuilder\Operations\IsNotNull;
@@ -74,10 +76,27 @@ class QueryBuilderTest extends TestCase
             ->andWhere(new Gt('u.version', 0))
             ->orderBy(['id' => 'DESC', 'userName'])
             ->groupBy(['userName', 'id' => 'DESC'])
+            ->join('test', 't1', new Eq('t1.id', 'u.id'))
             ->having(new Gte('COUNT(visit)', 1));
 
         self::assertEquals(
-            'SELECT id `f_0`, name `userName`, version `f_1` FROM users `u` WHERE u.id IS NOT NULL AND u.version > 0 GROUP BY userName ASC, id DESC HAVING COUNT(visit) >= 1 ORDER BY id DESC, userName ASC LIMIT 42, 100',
+            'SELECT id `f_0`, name `userName`, version `f_1` FROM users `u` JOIN test `t1` ON t1.id = u.id WHERE u.id IS NOT NULL AND u.version > 0 GROUP BY userName ASC, id DESC HAVING COUNT(visit) >= 1 ORDER BY id DESC, userName ASC LIMIT 42, 100',
+            $queryBuilder->getQuery()
+        );
+    }
+
+    public function testJoints(): void
+    {
+        $queryBuilder = new QueryBuilder();
+        $queryBuilder->select()
+            ->from('tests')
+            ->join('t1', 't1', new Eq(1, 1))
+            ->leftJoin('t1', 't1', new Eq(1, 1))
+            ->rightJoin('t1', 't1', new Eq(1, 1))
+            ->innerJoin('t1', 't1', new Eq(1, 1));
+
+        self::assertEquals(
+            'SELECT * FROM tests `t` JOIN t1 `t1` ON 1 = 1 LEFT JOIN t1 `t1` ON 1 = 1 RIGHT JOIN t1 `t1` ON 1 = 1 INNER JOIN t1 `t1` ON 1 = 1',
             $queryBuilder->getQuery()
         );
     }